refactor(frontend): add component domain folders with compatibility exports
