+++ frontend/components/data-display/index.ts
@@ -1,0 +1,3 @@
+export { default as EvidenceTable } from "./EvidenceTable";
+export { EVIDENCE_TYPES } from "./EvidenceTable";
+export { default as Timeline } from "./Timeline";
