+++ frontend/components/layout/index.ts
@@ -1,0 +1,2 @@
+export { default as MainLayout } from "./MainLayout";
+export { default as AdminLayout } from "./AdminLayout";
