+++ frontend/components/reports/index.ts
@@ -1,0 +1,2 @@
+export { default as IncidentSummaryCards } from "./IncidentSummaryCards";
+export { default as OverdueFollowUpList } from "./OverdueFollowUpList";
