+++ frontend/components/support/index.ts
@@ -1,0 +1,4 @@
+export { default as AlertsPanel } from "./AlertsPanel";
+export { default as CaseNotesPanel } from "./CaseNotesPanel";
+export { default as CaseTasksPanel } from "./CaseTasksPanel";
+export { default as MissingItemsPanel } from "./MissingItemsPanel";
